Split native source audit quality gate
